pyproject.toml
# ============================================================
# pyproject.toml
# Project dependencies managed with uv.
# Install: uv pip install -r pyproject.toml
# ============================================================

[project]
name = "chest-xray-classifier"
version = "1.0.0"
description = "Chest X-Ray Classification API using ResNet50"
requires-python = ">=3.11"

dependencies = [
    # API
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9", # required for FastAPI file uploads
    # ML
    "torch>=2.3.0",
    "torchvision>=0.18.0",
    "numpy>=1.26.0",
    "Pillow>=10.3.0",
    # Evaluation
    "scikit-learn>=1.5.0",
    "matplotlib>=3.9.0",
    "seaborn>=0.13.0",
    # ONNX export
    "onnx>=1.16.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",              # for FastAPI test client
]
